Changes to project structure updated classes to export default
