+++ room-ease/src/App.js
@@ -1,9 +1,11 @@
 import React, { Component } from 'react';
 import * as classes from './App.module.css';
 import { Route, Switch } from 'react-router-dom'
-import Login from './components/Account/Login/Login';
-import Signup from './components/Account/Signup/Signup';
-import JoinRoom from './components/Account/JoinScreen/JoinRoom';
+import Login from './containers/Account/Login/Login';
+import Signup from './containers/Account/Signup/Signup';
+import JoinRoom from './containers/JoinRoom/JoinRoom'
+import Homepage from './containers/Homepage/Homepage';
+
 class App extends Component {
 
   render() {
@@ -12,8 +14,8 @@ class App extends Component {
         <Switch>
           <Route path="/login" component={Login} />
           <Route path="/signup" component={Signup} />
-          <Route path="/join" component={JoinRoom}/>
-          <Route path="/" component={Login} />
+          <Route path="/room/join" component={JoinRoom} />
+          <Route path="/" component={Homepage} />
         </Switch>
       </div>
     );
