--- room-ease/src/components/Account/JoinScreen/joinroom.module.css
@@ -1,17 +1,0 @@
-.Form{
-    background-color: whitesmoke;    
-    margin-top: 10%;        
-    min-width: 450px;
-    display: inline-block;
-    text-align: center;
-    box-shadow: 3px 3px 5px black;   
-    border-radius: 5px;
-}
-.signin{
-    text-align: center;
-    font-family: 'Arial';
-    font-size: 1.6rem;
-}
-.error{
-    color: red;
-}
