+++ room-ease/src/containers/Account/Signup/signup.module.css
@@ -1,0 +1,17 @@
+.Form{
+    background-color: whitesmoke;    
+    margin-top: 10%;        
+    min-width: 450px;
+    display: inline-block;
+    text-align: center;
+    box-shadow: 3px 3px 5px black;   
+    border-radius: 5px;
+}
+.signup{
+    text-align: center;
+    font-family: 'Arial';
+    font-size: 1.6rem;
+}
+.error{
+    color: red;
+}
