+++ room-ease/src/containers/JoinRoom/JoinRoom.js
@@ -1,35 +1,37 @@
-import React, {Component} from 'react';
-import { Link } from "react-router-dom";
+import React, { Component } from 'react';
+
 import RegularButton from "../../components/inputs/RegularButton";
 import TextInput from "../../components/inputs/TextInput";
 import * as classes from './joinroom.module.css';
 import axios from 'axios';
-import {JOIN_ROOM_URL} from '../../../constants/ServerRoutes';
-export default class JoinRoom extends Component {
+import { JOIN_ROOM_URL } from '../../constants/ServerRoutes';
+
+
+class JoinRoom extends Component {
     constructor(props) {
         super(props);
         this.state = {
             roomName: '',
-            userID:'',
-            error:''
+            userID: '',
+            error: ''
         };
     }
-    handleRoomName= event=>{
-        this.setState({roomName:event.target.value});
+    handleRoomName = event => {
+        this.setState({ roomName: event.target.value });
     }
-    authenticateRoom= ()=>{
-        let loginCredentials = {userID:this.state.userID,roomName: this.state.roomName}
+    authenticateRoom = async () => {
+        let loginCredentials = { userID: this.state.userID, roomName: this.state.roomName }
         console.log(loginCredentials)
         let loginStatus = (await axios.post(JOIN_ROOM_URL, loginCredentials)).data;
         console.log(loginStatus)
         let { Result, Error } = loginStatus;
     }
-    render(){
-        let errorMessage= null;
-        if(this.state.error)
-            errorMessage=(<p className={classes.error}>{this.state.error}</p>)
+    render() {
+        let errorMessage = null;
+        if (this.state.error)
+            errorMessage = (<p className={classes.error}>{this.state.error}</p>)
 
-        return(
+        return (
             <div>
                 <h1>RoomEase</h1>
                 <div className={classes.div}>
@@ -37,7 +39,7 @@ export default class JoinRoom extends Component {
                     <TextInput hint="Enter room name" type="text" onChange={this.handleRoomName} ></TextInput>
                     <RegularButton disabled={this.state.roomName} text="Join" onClick={this.authenticateRoom}></RegularButton>
                     <h3> OR</h3>
-                    <RegularButton  text="Creat a room" onClick={this.createRoom}></RegularButton>
+                    <RegularButton text="Creat a room" onClick={this.createRoom}></RegularButton>
                 </div>
             </div>
         );
@@ -45,4 +47,7 @@ export default class JoinRoom extends Component {
 
     }
 }
+
+
+export default JoinRoom;
